requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-dateutil>=2.8.2
faker>=20.0.0
pydantic>=2.4.0
requests>=2.31.0
